requirements.txt
gspread
python-telegram-bot